fix server tests
